requirements.txt
requests
load_dotenv
fastapi
uvicorn
psycopg2
sqlalchemy
sqlmodel
alembic